Made a create a trip activity with error handling. Appears to be functional. Successfully adds it to the database for the home page to retrieve.
