+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,9 @@
         android:theme="@style/Theme.Group17HelloWorld"
         tools:targetApi="31">
         <activity
+            android:name=".CreateATripActivity"
+            android:exported="false" />
+        <activity
             android:name=".ViewTripDetails"
             android:exported="false" />
         <activity
